audio moved to separate thread
